requirements.txt
mysql-connector-python
watchdog
requests
PySide6 # Escolhendo PySide6 (Qt) para a GUI cross-platform

